refactor: simple identation pages
